+++ src/renderer/components/wallet/txs/send/SendFormCOSMOS.tsx
@@ -139,7 +139,7 @@ export const SendFormCOSMOS: React.FC<Props> = (props): JSX.Element => {
     let memoValue = form.getFieldValue('memo') as string
 
     // Check if a swap memo is detected
-    if (H.checkMemo(memoValue)) {
+    if (H.checkMemo(memoValue) && network === Network.Mainnet) {
       memoValue = H.memoCorrection(memoValue)
       setSwapMemoDetected(true)
 
@@ -151,7 +151,7 @@ export const SendFormCOSMOS: React.FC<Props> = (props): JSX.Element => {
 
     // Update the state with the adjusted memo value
     setCurrentMemo(memoValue)
-  }, [form, intl])
+  }, [form, intl, network])
 
   const oFee: O.Option<BaseAmount> = useMemo(() => FP.pipe(feeRD, RD.toOption), [feeRD])