+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - trade asset memo thornode error & swap input issue [#498](https://github.com/asgardex/asgardex-desktop/pull/498)
 - remove affiliate and bps for stagenet swaps [#500](https://github.com/asgardex/asgardex-desktop/pull/500)
 - fix earn disabled submit button [#481] (https://github.com/asgardex/asgardex-desktop/pull/501)
+- For stagenet don't apply affiliate or thorname [#512] (https://github.com/asgardex/asgardex-desktop/pull/501)
+
 # 1.22.8 (2024-11-05)
 
 ## add